+++ operator/src/main/java/oracle/kubernetes/operator/Main.java
@@ -35,12 +35,10 @@ import oracle.kubernetes.operator.helpers.CallBuilder;
 import oracle.kubernetes.operator.helpers.CallBuilderFactory;
 import oracle.kubernetes.operator.helpers.ConfigMapHelper;
 import oracle.kubernetes.operator.helpers.DomainPresenceInfo;
-import oracle.kubernetes.operator.helpers.DomainPresenceInfoManager;
 import oracle.kubernetes.operator.helpers.HealthCheckHelper;
 import oracle.kubernetes.operator.helpers.HealthCheckHelper.KubernetesVersion;
 import oracle.kubernetes.operator.helpers.ResponseStep;
 import oracle.kubernetes.operator.helpers.ServerKubernetesObjects;
-import oracle.kubernetes.operator.helpers.ServerKubernetesObjectsManager;
 import oracle.kubernetes.operator.logging.LoggingFacade;
 import oracle.kubernetes.operator.logging.LoggingFactory;
 import oracle.kubernetes.operator.logging.MessageKeys;
@@ -67,7 +65,7 @@ public class Main {
 
   private static final LoggingFacade LOGGER = LoggingFactory.getLogger("Operator", "Operator");
 
-  private static final String NS_STARTING_NOW = "NS_STARTING_NOW";
+  private static final String DPI_MAP = "DPI_MAP";
 
   static final TuningParameters tuningAndConfig;
 
@@ -86,6 +84,7 @@ public class Main {
   private static final Container container = new Container();
   private static final ScheduledExecutorService wrappedExecutorService =
       Engine.wrappedExecutorService("operator", container);
+  private static final DomainProcessor processor = DomainProcessor.getInstance();
 
   static {
     container
@@ -124,14 +123,6 @@ public class Main {
 
   static final String READINESS_PROBE_FAILURE_EVENT_FILTER =
       "reason=Unhealthy,type=Warning,involvedObject.fieldPath=spec.containers{weblogic-server}";
-
-  static Map<String, DomainPresenceInfo> getDomainPresenceInfos(String ns) {
-    return DomainPresenceInfoManager.getDomainPresenceInfos(ns);
-  }
-
-  static Map<String, ServerKubernetesObjects> getKubernetesObjects(String ns) {
-    return ServerKubernetesObjectsManager.getServerKubernetesObjects(ns);
-  }
 
   /**
    * Entry point
@@ -266,10 +257,7 @@ public class Main {
     @Override
     public NextAction apply(Packet packet) {
       AtomicBoolean a = isNamespaceStarted.computeIfAbsent(ns, (key) -> new AtomicBoolean(false));
-      boolean startingNow = !a.getAndSet(true);
-      packet.put(NS_STARTING_NOW, (Boolean) startingNow);
-
-      if (startingNow) {
+      if (!a.getAndSet(true)) {
         try {
           HealthCheckHelper.performSecurityChecks(version, operatorNamespace, ns);
         } catch (Throwable e) {
@@ -323,6 +311,7 @@ public class Main {
 
   static Step readExistingResources(String operatorNamespace, String ns) {
     return Step.chain(
+        new ReadExistingResourcesBeforeStep(),
         ConfigMapHelper.createScriptConfigMapStep(operatorNamespace, ns),
         createConfigMapStep(ns),
         readExistingPods(ns),
@@ -330,6 +319,15 @@ public class Main {
         readExistingServices(ns),
         readExistingIngresses(ns),
         readExistingDomains(ns));
+  }
+
+  private static class ReadExistingResourcesBeforeStep extends Step {
+    @SuppressWarnings("rawtypes")
+    @Override
+    public NextAction apply(Packet packet) {
+      packet.put(DPI_MAP, new ConcurrentHashMap());
+      return doNext(packet);
+    }
   }
 
   private static Step readExistingDomains(String ns) {
@@ -363,7 +361,7 @@ public class Main {
 
   private static ConfigMapAfterStep createConfigMapStep(String ns) {
     return new ConfigMapAfterStep(
-        ns, configMapWatchers, isNamespaceStopping(ns), DomainProcessor::dispatchConfigMapWatch);
+        ns, configMapWatchers, isNamespaceStopping(ns), processor::dispatchConfigMapWatch);
   }
 
   // -----------------------------------------------------------------------------
@@ -434,7 +432,7 @@ public class Main {
         ns,
         READINESS_PROBE_FAILURE_EVENT_FILTER,
         initialResourceVersion,
-        DomainProcessor::dispatchEventWatch,
+        processor::dispatchEventWatch,
         isNamespaceStopping(ns));
   }
 
@@ -443,7 +441,7 @@ public class Main {
         getThreadFactory(),
         ns,
         initialResourceVersion,
-        DomainProcessor::dispatchPodWatch,
+        processor::dispatchPodWatch,
         isNamespaceStopping(ns));
   }
 
@@ -452,7 +450,7 @@ public class Main {
         getThreadFactory(),
         ns,
         initialResourceVersion,
-        DomainProcessor::dispatchServiceWatch,
+        processor::dispatchServiceWatch,
         isNamespaceStopping(ns));
   }
 
@@ -461,7 +459,7 @@ public class Main {
         getThreadFactory(),
         ns,
         initialResourceVersion,
-        DomainProcessor::dispatchIngressWatch,
+        processor::dispatchIngressWatch,
         isNamespaceStopping(ns));
   }
 
@@ -495,18 +493,22 @@ public class Main {
 
     @Override
     public NextAction onSuccess(Packet packet, CallResponse<V1beta1IngressList> callResponse) {
+      @SuppressWarnings("unchecked")
+      Map<String, DomainPresenceInfo> dpis = (Map<String, DomainPresenceInfo>) packet.get(DPI_MAP);
+
       V1beta1IngressList result = callResponse.getResult();
       if (result != null) {
         for (V1beta1Ingress ingress : result.getItems()) {
           String domainUID = IngressWatcher.getIngressDomainUID(ingress);
           String clusterName = IngressWatcher.getIngressClusterName(ingress);
           if (domainUID != null && clusterName != null) {
-            DomainPresenceInfoManager.getOrCreate(ns, domainUID)
+            dpis.computeIfAbsent(domainUID, k -> new DomainPresenceInfo(ns, domainUID))
                 .getIngresses()
                 .put(clusterName, ingress);
           }
         }
       }
+
       if (!ingressWatchers.containsKey(ns)) {
         ingressWatchers.put(ns, createIngressWatcher(ns, getInitialResourceVersion(result)));
       }
@@ -534,35 +536,41 @@ public class Main {
 
     @Override
     public NextAction onSuccess(Packet packet, CallResponse<DomainList> callResponse) {
-      Boolean startingNow = (Boolean) packet.get(NS_STARTING_NOW);
-      if (startingNow == null) {
-        startingNow = Boolean.TRUE;
-      }
+      @SuppressWarnings("unchecked")
+      Map<String, DomainPresenceInfo> dpis = (Map<String, DomainPresenceInfo>) packet.get(DPI_MAP);
+
+      DomainProcessor x = packet.getSPI(DomainProcessor.class);
+      DomainProcessor dp = x != null ? x : processor;
 
       Set<String> domainUIDs = new HashSet<>();
       if (callResponse.getResult() != null) {
         for (Domain dom : callResponse.getResult().getItems()) {
           String domainUID = dom.getSpec().getDomainUID();
           domainUIDs.add(domainUID);
-          DomainPresenceInfo info = DomainPresenceInfoManager.getOrCreate(dom);
-          if (startingNow) {
-            // Update domain here if namespace is not yet running
-            info.setDomain(dom);
-          }
-          DomainProcessor.makeRightDomainPresence(info, domainUID, dom, true, false, false);
+          DomainPresenceInfo info =
+              dpis.compute(
+                  domainUID,
+                  (k, v) -> {
+                    if (v == null) {
+                      return new DomainPresenceInfo(dom);
+                    }
+                    v.setDomain(dom);
+                    return v;
+                  });
+          info.setPopulated(true);
+          dp.makeRightDomainPresence(info, true, false, false);
         }
       }
 
-      getDomainPresenceInfos(ns)
-          .forEach(
-              (key, value) -> {
-                if (!domainUIDs.contains(key)) {
-                  // This is a stranded DomainPresenceInfo.
-                  value.setDeleting(true);
-                  Domain dom = value.getDomain();
-                  DomainProcessor.makeRightDomainPresence(value, key, dom, true, true, false);
-                }
-              });
+      dpis.forEach(
+          (key, value) -> {
+            if (!domainUIDs.contains(key)) {
+              // This is a stranded DomainPresenceInfo.
+              value.setDeleting(true);
+              value.setPopulated(true);
+              dp.makeRightDomainPresence(value, true, true, false);
+            }
+          });
 
       if (!domainWatchers.containsKey(ns)) {
         domainWatchers.put(
@@ -580,7 +588,7 @@ public class Main {
           getThreadFactory(),
           ns,
           initialResourceVersion,
-          DomainProcessor::dispatchDomainWatch,
+          processor::dispatchDomainWatch,
           isNamespaceStopping(ns));
     }
   }
@@ -603,9 +611,9 @@ public class Main {
     public NextAction onSuccess(Packet packet, CallResponse<V1ServiceList> callResponse) {
       V1ServiceList result = callResponse.getResult();
 
-      Set<String> serviceNames = new HashSet<>();
-      Set<String> channelNames = new HashSet<>();
-      Set<String> clusterNames = new HashSet<>();
+      @SuppressWarnings("unchecked")
+      Map<String, DomainPresenceInfo> dpis = (Map<String, DomainPresenceInfo>) packet.get(DPI_MAP);
+
       if (result != null) {
         for (V1Service service : result.getItems()) {
           String domainUID = ServiceWatcher.getServiceDomainUID(service);
@@ -613,54 +621,22 @@ public class Main {
           String channelName = ServiceWatcher.getServiceChannelName(service);
           String clusterName = ServiceWatcher.getServiceClusterName(service);
           if (domainUID != null) {
-            DomainPresenceInfo info = DomainPresenceInfoManager.getOrCreate(ns, domainUID);
+            DomainPresenceInfo info =
+                dpis.computeIfAbsent(domainUID, k -> new DomainPresenceInfo(ns, domainUID));
             if (clusterName != null) {
-              clusterNames.add(clusterName);
               info.getClusters().put(clusterName, service);
             } else if (serverName != null) {
               ServerKubernetesObjects sko =
-                  ServerKubernetesObjectsManager.getOrCreate(info, domainUID, serverName);
+                  info.getServers().computeIfAbsent(serverName, k -> new ServerKubernetesObjects());
               if (channelName != null) {
-                channelNames.add(channelName);
                 sko.getChannels().put(channelName, service);
               } else {
-                serviceNames.add(service.getMetadata().getName());
                 sko.getService().set(service);
               }
             }
           }
         }
       }
-
-      getDomainPresenceInfos(ns)
-          .forEach(
-              (key, value) -> {
-                ConcurrentMap<String, V1Service> map = value.getClusters();
-                map.forEach(
-                    (ckey, cvalue) -> {
-                      map.compute(
-                          ckey,
-                          (k, current) -> {
-                            return clusterNames.contains(ckey) ? current : null;
-                          });
-                    });
-              });
-      getKubernetesObjects(ns)
-          .forEach(
-              (key, value) -> {
-                if (!serviceNames.contains(key)) {
-                  value.getService().set(null);
-                }
-                ConcurrentMap<String, V1Service> map = value.getChannels();
-                map.forEach(
-                    (ckey, cvalue) -> {
-                      map.compute(
-                          ckey,
-                          (k, current) -> {
-                            return channelNames.contains(ckey) ? current : null;
-                          });
-                    });
-              });
 
       if (!serviceWatchers.containsKey(ns)) {
         serviceWatchers.put(ns, createServiceWatcher(ns, getInitialResourceVersion(result)));
@@ -721,28 +697,22 @@ public class Main {
     public NextAction onSuccess(Packet packet, CallResponse<V1PodList> callResponse) {
       V1PodList result = callResponse.getResult();
 
-      Set<String> podNames = new HashSet<>();
+      @SuppressWarnings("unchecked")
+      Map<String, DomainPresenceInfo> dpis = (Map<String, DomainPresenceInfo>) packet.get(DPI_MAP);
+
       if (result != null) {
         for (V1Pod pod : result.getItems()) {
           String domainUID = PodWatcher.getPodDomainUID(pod);
           String serverName = PodWatcher.getPodServerName(pod);
           if (domainUID != null && serverName != null) {
-            podNames.add(pod.getMetadata().getName());
-            DomainPresenceInfo info = DomainPresenceInfoManager.getOrCreate(ns, domainUID);
+            DomainPresenceInfo info =
+                dpis.computeIfAbsent(domainUID, k -> new DomainPresenceInfo(ns, domainUID));
             ServerKubernetesObjects sko =
-                ServerKubernetesObjectsManager.getOrCreate(info, domainUID, serverName);
+                info.getServers().computeIfAbsent(serverName, k -> new ServerKubernetesObjects());
             sko.getPod().set(pod);
           }
         }
       }
-
-      getKubernetesObjects(ns)
-          .forEach(
-              (key, value) -> {
-                if (!podNames.contains(key)) {
-                  value.getPod().set(null);
-                }
-              });
 
       if (!podWatchers.containsKey(ns)) {
         podWatchers.put(ns, createPodWatcher(ns, getInitialResourceVersion(result)));
