+++ operator/src/main/java/oracle/kubernetes/operator/TuningParametersImpl.java
@@ -50,8 +50,8 @@ public class TuningParametersImpl extends ConfigMapConsumer implements TuningPar
 
     MainTuning main =
         new MainTuning(
-            (int) readTuningParameter("domainPresenceFailureRetrySeconds", 30),
-            (int) readTuningParameter("domainPresenceRecheckIntervalSeconds", 300),
+            (int) readTuningParameter("domainPresenceFailureRetrySeconds", 10),
+            (int) readTuningParameter("domainPresenceRecheckIntervalSeconds", 60),
             (int) readTuningParameter("statusUpdateTimeoutSeconds", 10),
             (int) readTuningParameter("statusUpdateUnchangedCountToDelayStatusRecheck", 10),
             readTuningParameter("statusUpdateInitialShortDelay", 3),
