+++ operator/src/test/java/oracle/kubernetes/operator/DomainPresenceTest.java
@@ -24,6 +24,7 @@ import static org.hamcrest.junit.MatcherAssert.assertThat;
 
 import com.meterware.simplestub.Memento;
 import com.meterware.simplestub.StaticStubSupport;
+import com.meterware.simplestub.Stub;
 import io.kubernetes.client.models.V1ConfigMap;
 import io.kubernetes.client.models.V1Event;
 import io.kubernetes.client.models.V1EventList;
@@ -45,15 +46,13 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicBoolean;
 import oracle.kubernetes.TestUtils;
 import oracle.kubernetes.operator.builders.StubWatchFactory;
 import oracle.kubernetes.operator.helpers.AsyncCallTestSupport;
-import oracle.kubernetes.operator.helpers.DomainPresenceInfoManager;
+import oracle.kubernetes.operator.helpers.DomainPresenceInfo;
 import oracle.kubernetes.operator.helpers.LegalNames;
 import oracle.kubernetes.operator.helpers.ServerKubernetesObjects;
-import oracle.kubernetes.operator.helpers.ServerKubernetesObjectsManager;
 import oracle.kubernetes.operator.work.ThreadFactorySingleton;
 import oracle.kubernetes.weblogic.domain.v2.Domain;
 import oracle.kubernetes.weblogic.domain.v2.DomainList;
@@ -84,26 +83,17 @@ public class DomainPresenceTest extends ThreadFactoryTestBase {
 
   @Before
   public void setUp() throws Exception {
-    getDomainPresenceInfoMap().clear();
-
     mementos.add(TestUtils.silenceOperatorLogger());
-    mementos.add(
-        installStub(ServerKubernetesObjectsManager.class, "serverMap", new ConcurrentHashMap<>()));
     mementos.add(testSupport.installRequestStepFactory());
     mementos.add(ClientFactoryStub.install());
     mementos.add(StubWatchFactory.install());
     mementos.add(installStub(ThreadFactorySingleton.class, "INSTANCE", this));
     mementos.add(
-        installStub(DomainProcessor.class, "FIBER_GATE", testSupport.createFiberGateStub()));
+        installStub(DomainProcessorImpl.class, "FIBER_GATE", testSupport.createFiberGateStub()));
 
     getStartedVariable();
     isNamespaceStopping = getStoppingVariable();
     isNamespaceStopping.computeIfAbsent(NS, k -> new AtomicBoolean(true)).set(true);
-  }
-
-  private Map getDomainPresenceInfoMap() throws NoSuchFieldException {
-    Memento domains = StaticStubSupport.preserve(DomainPresenceInfoManager.class, "domains");
-    return domains.getOriginalValue();
   }
 
   private static Memento installStub(Class<?> containingClass, String fieldName, Object newValue)
@@ -131,11 +121,31 @@ public class DomainPresenceTest extends ThreadFactoryTestBase {
     testSupport.throwOnCompletionFailure();
   }
 
+  public abstract static class DomainProcessorStub implements DomainProcessor {
+    private final Map<String, DomainPresenceInfo> dpis = new HashMap<>();
+
+    public Map<String, DomainPresenceInfo> getDomainPresenceInfos() {
+      return dpis;
+    }
+
+    @Override
+    public void makeRightDomainPresence(
+        DomainPresenceInfo info,
+        boolean explicitRecheck,
+        boolean isDeleting,
+        boolean isWillInterrupt) {
+      dpis.put(info.getDomainUID(), info);
+    }
+  }
+
   @Test
   public void whenNoPreexistingDomains_createEmptyDomainPresenceInfoMap() {
+    DomainProcessorStub dp = Stub.createStub(DomainProcessorStub.class);
+    testSupport.addComponent("DP", DomainProcessor.class, dp);
+
     readExistingResources();
 
-    assertThat(Main.getDomainPresenceInfos(NS), is(anEmptyMap()));
+    assertThat(dp.getDomainPresenceInfos(), is(anEmptyMap()));
   }
 
   private void readExistingResources() {
@@ -148,10 +158,13 @@ public class DomainPresenceTest extends ThreadFactoryTestBase {
     addDomainResource(UID, NS);
     addIngressResource(UID, NS, "cluster1");
 
+    DomainProcessorStub dp = Stub.createStub(DomainProcessorStub.class);
+    testSupport.addComponent("DP", DomainProcessor.class, dp);
+
     readExistingResources();
 
     assertThat(
-        Main.getDomainPresenceInfos(NS),
+        dp.getDomainPresenceInfos(),
         hasValue(domain(UID).withNamespace(NS).withIngressForCluster("cluster1")));
   }
 
@@ -202,16 +215,20 @@ public class DomainPresenceTest extends ThreadFactoryTestBase {
     addDomainResource(UID, NS);
     V1Service serviceResource = addServiceResource(UID, NS, "admin", "channel1");
 
+    DomainProcessorStub dp = Stub.createStub(DomainProcessorStub.class);
+    testSupport.addComponent("DP", DomainProcessor.class, dp);
+
     readExistingResources();
 
     String serverName = "admin";
     assertThat(
-        getServerKubernetesObjects(UID, serverName).getChannels(),
+        getServerKubernetesObjects(dp, UID, serverName).getChannels(),
         hasEntry(equalTo("channel1"), sameInstance(serviceResource)));
   }
 
-  private ServerKubernetesObjects getServerKubernetesObjects(String uid, String serverName) {
-    return Main.getDomainPresenceInfos(NS).get(uid).getServers().get(serverName);
+  private ServerKubernetesObjects getServerKubernetesObjects(
+      DomainProcessorStub dp, String uid, String serverName) {
+    return dp.getDomainPresenceInfos().get(uid).getServers().get(serverName);
   }
 
   private V1Service addServiceResource(
@@ -262,10 +279,13 @@ public class DomainPresenceTest extends ThreadFactoryTestBase {
     addDomainResource(UID, NS);
     V1Service serviceResource = addServiceResource(UID, NS, "admin");
 
+    DomainProcessorStub dp = Stub.createStub(DomainProcessorStub.class);
+    testSupport.addComponent("DP", DomainProcessor.class, dp);
+
     readExistingResources();
 
     assertThat(
-        getServerKubernetesObjects(UID, "admin").getService().get(), equalTo(serviceResource));
+        getServerKubernetesObjects(dp, UID, "admin").getService().get(), equalTo(serviceResource));
   }
 
   private V1Service addServiceResource(String uid, String namespace, String serverName) {
@@ -283,9 +303,12 @@ public class DomainPresenceTest extends ThreadFactoryTestBase {
     addDomainResource(UID, NS);
     V1Pod podResource = addPodResource(UID, NS, "admin");
 
+    DomainProcessorStub dp = Stub.createStub(DomainProcessorStub.class);
+    testSupport.addComponent("DP", DomainProcessor.class, dp);
+
     readExistingResources();
 
-    assertThat(getServerKubernetesObjects(UID, "admin").getPod().get(), equalTo(podResource));
+    assertThat(getServerKubernetesObjects(dp, UID, "admin").getPod().get(), equalTo(podResource));
   }
 
   private V1Pod addPodResource(String uid, String namespace, String serverName) {
@@ -305,10 +328,13 @@ public class DomainPresenceTest extends ThreadFactoryTestBase {
     addPodResource(UID, NS, "admin");
     addEventResource(UID, "admin", READINESS_PROBE_NOT_READY_STATE + "do something!");
 
+    DomainProcessorStub dp = Stub.createStub(DomainProcessorStub.class);
+    testSupport.addComponent("DP", DomainProcessor.class, dp);
+
     readExistingResources();
 
     assertThat(
-        getServerKubernetesObjects(UID, "admin").getLastKnownStatus().get(),
+        getServerKubernetesObjects(dp, UID, "admin").getLastKnownStatus().get(),
         equalTo("do something!"));
   }
 
@@ -318,9 +344,13 @@ public class DomainPresenceTest extends ThreadFactoryTestBase {
     addPodResource(UID, NS, "admin");
     addEventResource(UID, "admin", "ignore this event");
 
+    DomainProcessorStub dp = Stub.createStub(DomainProcessorStub.class);
+    testSupport.addComponent("DP", DomainProcessor.class, dp);
+
     readExistingResources();
 
-    assertThat(getServerKubernetesObjects(UID, "admin").getLastKnownStatus().get(), nullValue());
+    assertThat(
+        getServerKubernetesObjects(dp, UID, "admin").getLastKnownStatus().get(), nullValue());
   }
 
   private void addEventResource(String uid, String serverName, String message) {
